detection-service/requirements.txt
torch>=1.9.0
torchvision>=0.10.0
opencv-python>=4.5.0
numpy>=1.21.0
pillow>=8.3.0
ultralytics>=8.0.0
flask>=2.0.0
flask-cors>=3.0.0
websockets>=10.0
python-socketio>=5.0.0
eventlet>=0.33.0
